Move restaurant filtering logic to provider
- Move restaurant filtering logic to centralized location
- Reduces prop drilling in App.tsx and RestaurantPage.tsx
